fix: レビュー指摘対応(M-2,m-1) Projection型修正とJavadoc修正 / Address review findings: fix Projection types and Javadoc
M-2: DailySummaryReportRowの数量系カラムをBIGINT対応のLong型に変更
m-1: ReturnsReportRepositoryのJavadocをクエリ実装と整合させる

+++ backend/src/main/java/com/wms/report/repository/projection/DailySummaryReportRow.java
@@ -17,19 +17,19 @@ public interface DailySummaryReportRow {
 
     Integer getInboundLineCount();
 
-    Integer getInboundQuantityTotal();
+    Long getInboundQuantityTotal();
 
     Integer getOutboundCount();
 
     Integer getOutboundLineCount();
 
-    Integer getOutboundQuantityTotal();
+    Long getOutboundQuantityTotal();
 
     Integer getReturnCount();
 
     Integer getReturnQuantityTotal();
 
-    Integer getInventoryQuantityTotal();
+    Long getInventoryQuantityTotal();
 
     Integer getUnreceivedCount();
